requirements.txt
anthropic
pdfplumber
playwright
pyyaml
requests
rich
pydantic
